refactor: fold _save_png helpers into the quicklook Plot base class
Now that quicklook has a base.py Plot class, the standalone _save_png module
is redundant. Move its two functions -- save_png and save_image_png -- onto
Plot as staticmethods (they use no instance state, matching the module's
convention for pure helpers like _norm_order/_snr) and delete _save_png.py.
The function bodies are unchanged; call sites in base/level0/level1 now go
through self.save_png / self.save_image_png.

Test: quicklook/plot regression suite (143 passed); ruff format + check clean.
