+++ kpfpipe/quality_control/quicklook/level1.py
@@ -6,7 +6,6 @@ import matplotlib.pyplot as plt
 import numpy as np
 
 from kpfpipe.modules.image_assembly import RN_KEYS
-from kpfpipe.quality_control.quicklook._save_png import save_image_png, save_png
 from kpfpipe.quality_control.quicklook.base import Plot
 
 
@@ -148,7 +147,7 @@ class PlotL1(Plot):
                     self.output_dir,
                     f"{self.obs_id}_L1_image_{chip.lower()}_full_res.png",
                 )
-                save_image_png(
+                self.save_image_png(
                     image,
                     fig_path,
                     cmap=cmap,
@@ -161,7 +160,7 @@ class PlotL1(Plot):
                     self.output_dir,
                     f"{self.obs_id}_L1_image_{chip.lower()}_zoomable.png",
                 )
-                save_png(fig, fig_path, dpi=600, compress_level=1)
+                self.save_png(fig, fig_path, dpi=600, compress_level=1)
 
         return fig
